feat: integrate run history mock data  [#1401]
